+++ CHANGES.md
@@ -10,6 +10,10 @@
 - Added support for `debugColorTiles` in `ModelExperimental`. [#10071](https://github.com/CesiumGS/cesium/pull/10071)
 - Added support for shadows in `ModelExperimental`. [#10077](https://github.com/CesiumGS/cesium/pull/10077)
 
+##### Fixes :wrench:
+
+- Fixed a bug where updating `ModelExperimental`'s model matrix would not update its bounding sphere. [#10078](https://github.com/CesiumGS/cesium/pull/10078)
+
 ### 1.90 - 2022-02-01
 
 ##### Additions :tada: